Yundun_chinese/requirements.txt
absl-py==0.9.0
albumentations==0.4.5
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
efficientnet-pytorch==0.6.3
Flask==1.1.1
future==0.18.2
google-auth==1.14.2
google-auth-oauthlib==0.4.1
grpcio==1.28.1
h5py==2.10.0
html5lib==0.9999999
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imgaug==0.2.6
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.16.0
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
threadpoolctl==2.0.0
tifffile==2020.5.11
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
